refactor: MainWindow.xaml.cs を partial class でファイル分割 (#101 フェーズ1)
コードの変更なし。2360行のモノリスを責務ごとに8ファイルへ分割。

- MainWindow.xaml.cs     フィールド・定数・コンストラクタ・共通ヘルパー (330行)
- MainWindow.Settings.cs  設定・プロファイル永続化・設定ダイアログ・Aboutダイアログ (507行)
- MainWindow.Theme.cs     テーマ・タイトルバー (69行)
- MainWindow.Profiles.cs  プロファイル管理・バッジ (194行)
- MainWindow.Post.cs      投稿・Compose・WebMessageReceived (155行)
- MainWindow.Timeline.cs  タイムライン CRUD・永続化・DnD (518行)
- MainWindow.HardReload.cs ハードリロード・自動アクティブタイマー (207行)
- MainWindow.Updates.cs   アップデート確認 (125行)
- MainWindow.WebView2.cs  WebView2初期化・拡張機能 (467行)
