+++ MainWindow.xaml.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Automation;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media;
@@ -308,1434 +308,6 @@ namespace XTimelineViewer
             _ = CheckForUpdatesInBackgroundAsync();
         }
 
-        // ── App settings ──────────────────────────────────────────────────────
-
-        private void LoadSettings()
-        {
-            try
-            {
-                var json = File.ReadAllText(SettingsFilePath);
-                _appSettings = JsonSerializer.Deserialize<AppSettings>(json) ?? new();
-            }
-            catch { /* ファイルが存在しない場合などは無視 */ }
-            _appSettings.SeparateComposeEnv = false; // 廃止予定: 強制無効化 (#17)
-        }
-
-        private void SaveSettings()
-        {
-            Directory.CreateDirectory(Path.GetDirectoryName(SettingsFilePath)!);
-            File.WriteAllText(SettingsFilePath, JsonSerializer.Serialize(_appSettings, JsonOptions));
-        }
-
-        private void LoadProfiles()
-        {
-            try
-            {
-                var json = File.ReadAllText(ProfilesFilePath);
-                _profiles = JsonSerializer.Deserialize<List<ProfileConfig>>(json) ?? [];
-            }
-            catch { /* ファイルが存在しない場合などは無視 */ }
-            if (_profiles.Count == 0)
-            {
-                _profiles.Add(new ProfileConfig { Id = "default", Name = "Default" });
-                SaveProfiles();
-            }
-        }
-
-        private void SaveProfiles()
-        {
-            Directory.CreateDirectory(Path.GetDirectoryName(ProfilesFilePath)!);
-            File.WriteAllText(ProfilesFilePath, JsonSerializer.Serialize(_profiles, JsonOptions));
-        }
-
-        private void CleanupOrphanedProfiles()
-        {
-            try
-            {
-                var dir = GetProfilesDataDir();
-                if (!Directory.Exists(dir)) return;
-                var knownIds = new HashSet<string>(_profiles.Select(p => p.Id));
-                foreach (var folder in Directory.GetDirectories(dir))
-                {
-                    var name = Path.GetFileName(folder);
-                    if (!knownIds.Contains(name))
-                    {
-                        try
-                        {
-                            Directory.Delete(folder, recursive: true);
-                            Debug.WriteLine($"[Profile] Cleaned up orphaned folder: {name}");
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug.WriteLine($"[Profile] Failed to clean up orphaned folder: {name} ({ex.Message})");
-                        }
-                    }
-                }
-            }
-            catch { }
-        }
-
-        private void RestartAutoActivateTimer()
-        {
-            if (_autoActivateTimer is null) return;
-            _autoActivateTimer.Stop();
-            _autoActivateStartTime = DateTimeOffset.Now;
-            _autoActivateTimer.Start();
-        }
-
-        private void ApplyAutoActivateTimer()
-        {
-            _autoActivateTimer?.Stop();
-            if (_appSettings.AutoActivateMinutes <= 0) return;
-
-            _autoActivateTimer = new DispatcherTimer
-            {
-                Interval = TimeSpan.FromMinutes(_appSettings.AutoActivateMinutes)
-            };
-            _autoActivateStartTime = DateTimeOffset.Now;
-            _autoActivateTimer.Tick += (_, _) =>
-            {
-                _autoActivateStartTime = DateTimeOffset.Now;
-                if (_pointerOverWebViews.Count > 0)                 return;  // ① ポインターオーバー中
-                if (_dialogOpenCount > 0)                           return;  // ② ダイアログ表示中
-                if (_focusedHeaderGrid is not null && _homeHeaderGrids.Contains(_focusedHeaderGrid))  return;  // ④ ホームにフォーカス中
-
-                foreach (var wv in _webViews)
-                {
-                    if (wv.CoreWebView2 is null) continue;
-                    if (!Uri.TryCreate(wv.CoreWebView2.Source, UriKind.Absolute, out var src)) continue;
-                    if (!src.AbsolutePath.StartsWith("/home", StringComparison.OrdinalIgnoreCase)) continue;
-                    if (_urlDivergedWebViews.Contains(wv)) continue;  // ③ 別ページ閲覧中
-
-                    if (_webViewToPane.TryGetValue(wv, out var pane) &&
-                        _paneToSetFocus.TryGetValue(pane, out var setFocus))
-                    {
-                        setFocus();
-                        break;
-                    }
-                }
-            };
-            _autoActivateTimer.Start();
-        }
-
-        [DllImport("dwmapi.dll")]
-        private static extern int DwmSetWindowAttribute(IntPtr hwnd, int attr, ref int attrValue, int attrSize);
-        private const int DWMWA_USE_IMMERSIVE_DARK_MODE = 20;
-
-        // Win32 タイトルバーは WinUI の RequestedTheme の影響を受けないため
-        // DWM 属性で直接ダークモードを指定する。子ウィンドウにも適用できるよう static に。
-        internal static void ApplyTitleBarTheme(Window window, ElementTheme theme)
-        {
-            var hwnd = WinRT.Interop.WindowNative.GetWindowHandle(window);
-            var dark = theme == ElementTheme.Dark ? 1
-                     : theme == ElementTheme.Light ? 0
-                     : (Application.Current.RequestedTheme == ApplicationTheme.Dark ? 1 : 0);
-            DwmSetWindowAttribute(hwnd, DWMWA_USE_IMMERSIVE_DARK_MODE, ref dark, sizeof(int));
-        }
-
-        private void ApplySavedTheme()
-        {
-            var theme = _appSettings.Theme switch
-            {
-                "Light" => ElementTheme.Light,
-                "Dark"  => ElementTheme.Dark,
-                _       => ElementTheme.Default,
-            };
-            ((FrameworkElement)Content).RequestedTheme = theme;
-            ApplyTitleBarTheme(this, theme);
-            ApplyThemeToWebViews();
-        }
-
-        private void ManageProfilesMenuItem_Click(object _, RoutedEventArgs __)
-        {
-            var ownerHwnd = WinRT.Interop.WindowNative.GetWindowHandle(this);
-            var win = new ManageProfilesWindow(_profiles, ProfileBadgeColors, ownerHwnd,
-                profileId => _configs.Count(c => c.ProfileId == profileId));
-            var childTheme = ((FrameworkElement)Content).RequestedTheme;
-            ((FrameworkElement)win.Content).RequestedTheme = childTheme;
-            ApplyTitleBarTheme(win, childTheme);
-            win.ProfilesChanged += (__, args) =>
-            {
-                foreach (var change in args.Changes)
-                {
-                    var p = _profiles.FirstOrDefault(p => p.Id == change.ProfileId);
-                    if (p == null) continue;
-                    p.Name = change.NewName;
-                    p.BadgeColorIndex = change.NewColorIndex;
-                    p.BadgeText = change.NewBadgeText;
-                }
-                SaveProfiles();
-                RefreshAllProfileBadges();
-            };
-            win.ProfileCreated += (__, profile) =>
-            {
-                SaveProfiles();
-                RefreshAllProfileBadges();
-                Debug.WriteLine($"[Profile] Saved to profiles.json: Id={profile.Id}, Name={profile.Name}");
-            };
-            win.ProfileDeleteRequested += async (__, profileId) =>
-            {
-                RemoveTimelinesForProfile(profileId);
-                _profileEnvs.Remove(profileId);
-                var profile = _profiles.FirstOrDefault(p => p.Id == profileId);
-                if (profile != null) _profiles.Remove(profile);
-                if (_profiles.Count == 0)
-                    _profiles.Add(new ProfileConfig { Id = "default", Name = "Default" });
-                SaveProfiles();
-                try
-                {
-                    var folder = Path.Combine(GetProfilesDataDir(), profileId);
-                    if (Directory.Exists(folder))
-                        Directory.Delete(folder, recursive: true);
-                }
-                catch (Exception ex)
-                {
-                    Debug.WriteLine($"[Profile] Failed to delete profile folder: {ex.Message}");
-                }
-                await SaveTimelinesAsync();
-                RefreshAllProfileBadges();
-                Debug.WriteLine($"[Profile] Deleted: {profileId}");
-            };
-            win.Activate();
-        }
-
-        private void RefreshAllProfileBadges()
-        {
-            for (int i = 0; i < _configs.Count && i < TimelinePanel.Children.Count; i++)
-            {
-                var pane = (Grid)TimelinePanel.Children[i];
-                var headerGrid = pane.Children.OfType<Grid>().FirstOrDefault();
-                if (headerGrid == null) continue;
-                var oldBadge = headerGrid.Children.OfType<Border>()
-                    .FirstOrDefault(b => Grid.GetColumn(b) == 1);
-                if (oldBadge != null)
-                    headerGrid.Children.Remove(oldBadge);
-                var newBadge = CreateProfileBadge(_configs[i].ProfileId);
-                Grid.SetColumn(newBadge, 1);
-                headerGrid.Children.Add(newBadge);
-            }
-        }
-
-        private void RemoveTimelinesForProfile(string profileId)
-        {
-            var indices = new List<int>();
-            for (int i = 0; i < _configs.Count; i++)
-                if (_configs[i].ProfileId == profileId)
-                    indices.Add(i);
-
-            for (int i = indices.Count - 1; i >= 0; i--)
-            {
-                var idx = indices[i];
-                if (idx >= TimelinePanel.Children.Count) continue;
-                var pane = (Grid)TimelinePanel.Children[idx];
-                var wv = pane.Children.OfType<WebView2>().FirstOrDefault();
-                if (wv != null)
-                {
-                    CleanupWebView(wv);
-                }
-                var headerGrid = pane.Children.OfType<Grid>().FirstOrDefault();
-                if (headerGrid != null)
-                {
-                    _homeHeaderGrids.Remove(headerGrid);
-                    if (_focusedHeaderGrid == headerGrid)
-                        _focusedHeaderGrid = null;
-                }
-                _paneToSetFocus.Remove(pane);
-                TimelinePanel.Children.RemoveAt(idx);
-                _configs.RemoveAt(idx);
-            }
-
-            if (_hardReloadUiUpdaters.Count == 0)
-            {
-                _hardReloadUiTimer?.Stop();
-                _hardReloadUiTimer = null;
-            }
-            if (_focusedHeaderGrid == null)
-                foreach (var r in _headerRefreshers) r();
-
-            if (TimelinePanel.Children.Count == 0)
-            {
-                TimelineScroll.Visibility = Visibility.Collapsed;
-                DropHintBorder.Visibility = Visibility.Visible;
-            }
-        }
-
-        private async void AppSettingsMenuItem_Click(object _, RoutedEventArgs __)
-        {
-            var themeCombo = new ComboBox
-            {
-                ItemsSource   = new List<string> { R.Get("Theme_System"), R.Get("Theme_Light"), R.Get("Theme_Dark") },
-                SelectedIndex = _appSettings.Theme switch { "Light" => 1, "Dark" => 2, _ => 0 },
-                MinWidth      = 140
-            };
-
-            var langValues = new[] { "system", "ja-JP", "en-US" };
-            var langIdx    = Array.IndexOf(langValues, _appSettings.Language);
-            var langCombo  = new ComboBox
-            {
-                ItemsSource   = new List<string> { R.Get("Language_System"), R.Get("Language_JA"), R.Get("Language_EN") },
-                SelectedIndex = langIdx < 0 ? 0 : langIdx,
-                MinWidth      = 140
-            };
-
-            var openFolderBtn = new Button { Content = R.Get("Button_OpenFolder") };
-            openFolderBtn.Click += async (_, _) =>
-            {
-                var folder = Path.GetDirectoryName(SettingsFilePath)!;
-                Directory.CreateDirectory(folder);
-                await Windows.System.Launcher.LaunchFolderPathAsync(folder);
-            };
-
-            static Grid MakeRow(string label, FrameworkElement control, Thickness? margin = null)
-            {
-                var g = new Grid { Margin = margin ?? new Thickness(0, 6, 0, 0) };
-                g.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });
-                g.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-                var lbl = new TextBlock
-                {
-                    Text = label,
-                    VerticalAlignment = VerticalAlignment.Center,
-                    TextWrapping = TextWrapping.Wrap
-                };
-                Grid.SetColumn(lbl, 0);
-                Grid.SetColumn(control, 1);
-                g.Children.Add(lbl);
-                g.Children.Add(control);
-                return g;
-            }
-
-            var panel = new StackPanel { MinWidth = 400 };
-            panel.Children.Add(MakeRow(R.Get("Settings_Theme"), themeCombo, new Thickness(0)));
-            panel.Children.Add(MakeRow(R.Get("Settings_Language"), langCombo));
-            panel.Children.Add(MakeRow(R.Get("Settings_ExportFolder"), openFolderBtn));
-            panel.Children.Add(new TextBlock
-            {
-                Text                   = Path.GetDirectoryName(SettingsFilePath),
-                TextWrapping           = TextWrapping.Wrap,
-                IsTextSelectionEnabled = true,
-                Opacity                = 0.6,
-                FontSize               = 12,
-                Margin                 = new Thickness(0, 2, 0, 0),
-            });
-            panel.Children.Add(new NavigationViewItemSeparator { Margin = new Thickness(0, 12, 0, 8) });
-            panel.Children.Add(new TextBlock
-            {
-                Text       = R.Get("Section_Experimental"),
-                FontWeight = Microsoft.UI.Text.FontWeights.SemiBold
-            });
-            var separateEnvToggle = new ToggleSwitch
-            {
-                IsOn              = false,
-                IsEnabled         = false, // 廃止予定 (#17)
-                OnContent         = R.Get("Toggle_On"),
-                OffContent        = R.Get("Toggle_Off"),
-                Margin              = new Thickness(12, 0, 0, 0),
-                VerticalAlignment   = VerticalAlignment.Center,
-                HorizontalAlignment = HorizontalAlignment.Right
-            };
-            panel.Children.Add(MakeRow(R.Get("Settings_SeparateCompose"), separateEnvToggle));
-
-            var openPostToggle = new ToggleSwitch
-            {
-                IsOn                = _appSettings.OpenComposerInBrowser,
-                OnContent           = R.Get("Toggle_On"),
-                OffContent          = R.Get("Toggle_Off"),
-                Margin              = new Thickness(12, 0, 0, 0),
-                VerticalAlignment   = VerticalAlignment.Center,
-                HorizontalAlignment = HorizontalAlignment.Right
-            };
-            panel.Children.Add(MakeRow(R.Get("Settings_OpenComposerInBrowser"), openPostToggle));
-
-            var openTimestampToggle = new ToggleSwitch
-            {
-                IsOn                = _appSettings.OpenTimestampInBrowser,
-                OnContent           = R.Get("Toggle_On"),
-                OffContent          = R.Get("Toggle_Off"),
-                Margin              = new Thickness(12, 0, 0, 0),
-                VerticalAlignment   = VerticalAlignment.Center,
-                HorizontalAlignment = HorizontalAlignment.Right
-            };
-            panel.Children.Add(MakeRow(R.Get("Settings_OpenTimestampInBrowser"), openTimestampToggle));
-
-            var autoActivateBox = new NumberBox
-            {
-                Value                   = _appSettings.AutoActivateMinutes,
-                Minimum                 = 0,
-                Maximum                 = 60,
-                SmallChange             = 1,
-                LargeChange             = 5,
-                SpinButtonPlacementMode = NumberBoxSpinButtonPlacementMode.Inline,
-                Width                   = 160,
-            };
-            panel.Children.Add(MakeRow(R.Get("Settings_AutoActivate"), autoActivateBox));
-
-            var showAutoActivateLabelToggle = new ToggleSwitch
-            {
-                IsOn                = _appSettings.ShowAutoActivateLabel,
-                OnContent           = R.Get("Toggle_On"),
-                OffContent          = R.Get("Toggle_Off"),
-                Margin              = new Thickness(12, 0, 0, 0),
-                VerticalAlignment   = VerticalAlignment.Center,
-                HorizontalAlignment = HorizontalAlignment.Right
-            };
-            panel.Children.Add(MakeRow(R.Get("Settings_ShowAutoActivateLabel"), showAutoActivateLabelToggle));
-
-            var dlg = new ContentDialog
-            {
-                Title             = R.Get("AppSettings_Title"),
-                Content           = panel,
-                PrimaryButtonText = R.Get("Button_Save"),
-                CloseButtonText   = R.Get("Button_Cancel"),
-                DefaultButton     = ContentDialogButton.Primary,
-                XamlRoot          = Content.XamlRoot,
-                RequestedTheme    = ((FrameworkElement)Content).ActualTheme
-            };
-
-                        if (await ShowDialogAsync(dlg) == ContentDialogResult.Primary)
-            {
-                _appSettings.Theme = themeCombo.SelectedIndex switch { 1 => "Light", 2 => "Dark", _ => "Default" };
-                _appSettings.OpenComposerInBrowser = openPostToggle.IsOn;
-                _appSettings.OpenTimestampInBrowser = openTimestampToggle.IsOn;
-                _appSettings.AutoActivateMinutes   = (int)Math.Clamp(autoActivateBox.Value, 0, 60);
-                _appSettings.ShowAutoActivateLabel = showAutoActivateLabelToggle.IsOn;
-
-                var newLang    = langValues[Math.Max(0, Math.Min(langCombo.SelectedIndex, langValues.Length - 1))];
-                var langChanged = newLang != _appSettings.Language;
-                _appSettings.Language = newLang;
-
-                SaveSettings();
-                ApplySavedTheme();
-
-                var tsFlag = _appSettings.OpenTimestampInBrowser ? "true" : "false";
-                foreach (var wv in _webViews)
-                    if (wv.CoreWebView2 is not null)
-                        await wv.CoreWebView2.ExecuteScriptAsync(
-                            $"window._xtvOpenTimestampInBrowser = {tsFlag};");
-
-                ApplyAutoActivateTimer();
-
-                if (langChanged)
-                {
-                    var notifDlg = new ContentDialog
-                    {
-                        Content         = R.Get("RestartRequired"),
-                        CloseButtonText = R.Get("Button_Close"),
-                        XamlRoot        = Content.XamlRoot,
-                        RequestedTheme  = ((FrameworkElement)Content).ActualTheme
-                    };
-                    await ShowDialogAsync(notifDlg);
-                }
-            }
-        }
-
-        private async void AboutMenuItem_Click(object _, RoutedEventArgs __)
-        {
-            var currentVersion = Assembly.GetExecutingAssembly().GetName().Version!;
-            var versionStr     = currentVersion.ToString(3);
-
-            var edgeChannel = FindEdgeDevVersionFolder() is not null
-                ? R.Get("EdgeChannel_Dev")
-                : R.Get("EdgeChannel_Runtime");
-            string edgeVersion;
-            try
-            {
-                edgeVersion = CoreWebView2Environment.GetAvailableBrowserVersionString(
-                    FindEdgeDevVersionFolder() ?? "");
-            }
-            catch
-            {
-                edgeVersion = _profileEnvs.Values.FirstOrDefault()?.BrowserVersionString
-                              ?? R.Get("Version_Unknown");
-            }
-            var versionInfoText = $"XTimelineViewer v{versionStr}\r\n{edgeChannel} {edgeVersion}";
-
-            var issueBody = Uri.EscapeDataString(
-                $"- {R.Get("IssueLabel_AppVersion")}: v{versionStr}\n" +
-                $"- {R.Get("IssueLabel_EdgeVersion")}: {edgeChannel} {edgeVersion}\n" +
-                $"- {R.Get("IssueLabel_Symptoms")}:\n");
-            var issueUrl    = $"https://github.com/daruyanagi/XTimelineViewer/issues/new?labels=bug&title=&body={issueBody}";
-            var repoUrl     = "https://github.com/daruyanagi/XTimelineViewer";
-            var fallbackUrl = repoUrl + "/releases/latest";
-
-            // ── Header ────────────────────────────────────────────────────────
-            var iconPath = Path.Combine(AppContext.BaseDirectory, "Assets", "StoreLogo.png");
-
-            var textStack = new StackPanel { Spacing = 3, VerticalAlignment = VerticalAlignment.Center };
-            textStack.Children.Add(new TextBlock
-            {
-                Text       = "XTimelineViewer",
-                FontSize   = 20,
-                FontWeight = Microsoft.UI.Text.FontWeights.SemiBold,
-            });
-            textStack.Children.Add(new TextBlock { Text = $"v{versionStr}", FontSize = 13, Opacity = 0.7 });
-            textStack.Children.Add(new TextBlock { Text = R.Get("About_Copyright"), FontSize = 12, Opacity = 0.6 });
-
-            var titleRow = new StackPanel
-            {
-                Orientation = Orientation.Horizontal,
-                Spacing     = 12,
-                Margin      = new Thickness(0, 0, 0, 8),
-            };
-            if (File.Exists(iconPath))
-                titleRow.Children.Add(new Microsoft.UI.Xaml.Controls.Image
-                {
-                    Source            = new Microsoft.UI.Xaml.Media.Imaging.BitmapImage(new Uri(iconPath)),
-                    Width             = 48,
-                    Height            = 48,
-                    VerticalAlignment = VerticalAlignment.Top,
-                });
-            titleRow.Children.Add(textStack);
-
-            var copyBtn = new Button
-            {
-                Content = new StackPanel
-                {
-                    Orientation = Orientation.Horizontal,
-                    Spacing     = 6,
-                    Children    =
-                    {
-                        new FontIcon { Glyph = "", FontFamily = new FontFamily("Segoe Fluent Icons"), FontSize = 14 },
-                        new TextBlock { Text = R.Get("Button_Copy") },
-                    }
-                },
-                Margin = new Thickness(0, 4, 8, 0),
-            };
-            copyBtn.Click += (_, _) =>
-            {
-                var dp = new DataPackage();
-                dp.SetText(versionInfoText);
-                Clipboard.SetContent(dp);
-            };
-
-            // ── Update check ─────────────────────────────────────────────────
-            string? latestReleaseUrl = null;
-
-            var statusText = new TextBlock { FontSize = 13, Margin = new Thickness(0, 6, 0, 0), Visibility = Visibility.Collapsed };
-
-            bool hasWinget = !PackageContext.IsPackaged && FindWinget() is not null;
-            var updateBtnLabel = PackageContext.IsPackaged
-                ? R.Get("CheckUpdate_Download_Store")
-                : hasWinget
-                    ? R.Get("CheckUpdate_Download_Winget")
-                    : R.Get("CheckUpdate_Download_GitHub");
-            var updateBtn = new Button { Content = updateBtnLabel, Margin = new Thickness(0, 4, 0, 0), Visibility = Visibility.Collapsed };
-
-            // キャッシュがあれば初期表示
-            if (_appSettings.CachedLatestVersion is { } cached
-                && Version.TryParse(cached.TrimStart('v'), out var cachedVersion)
-                && IsUpdateAvailable(currentVersion, cachedVersion))
-            {
-                latestReleaseUrl = $"{repoUrl}/releases/tag/{cached}";
-                statusText.Text       = string.Format(R.Get("CheckUpdate_Available"), cached);
-                statusText.Visibility = Visibility.Visible;
-                updateBtn.Visibility  = Visibility.Visible;
-            }
-
-            var checkBtn = new Button { Content = R.Get("CheckUpdate_Btn"), Margin = new Thickness(0, 4, 0, 0) };
-            checkBtn.Click += async (_, _) =>
-            {
-                checkBtn.IsEnabled    = false;
-                statusText.Text       = R.Get("CheckUpdate_Checking");
-                statusText.Visibility = Visibility.Visible;
-                updateBtn.Visibility  = Visibility.Collapsed;
-                try
-                {
-                    var (latest, tag, freshUrl) = await FetchLatestReleaseAsync();
-                    latestReleaseUrl = freshUrl;
-                    _appSettings.LastUpdateCheck = DateTime.UtcNow.ToString("O");
-                    if (IsUpdateAvailable(currentVersion, latest))
-                    {
-                        _appSettings.CachedLatestVersion = tag;
-                        statusText.Text      = string.Format(R.Get("CheckUpdate_Available"), tag);
-                        updateBtn.Visibility = Visibility.Visible;
-                    }
-                    else
-                    {
-                        _appSettings.CachedLatestVersion = null;
-                        statusText.Text = R.Get("CheckUpdate_Latest");
-                    }
-                    SaveSettings();
-                    UpdateMenuUpdateBadge();
-                }
-                catch
-                {
-                    statusText.Text = R.Get("CheckUpdate_Error");
-                }
-                finally
-                {
-                    checkBtn.IsEnabled = true;
-                }
-            };
-
-            ContentDialog dlg = null!;
-
-            updateBtn.Click += async (_, _) =>
-            {
-                var url = latestReleaseUrl ?? fallbackUrl;
-                if (PackageContext.IsPackaged)
-                {
-                    _ = Windows.System.Launcher.LaunchUriAsync(new Uri("ms-windows-store://pdp/?ProductId=9P308HB5BLJ1"));
-                }
-                else if (FindWinget() is not null)
-                {
-                    dlg.Hide();
-                    await SelfUpdateViaWingetAsync(url);
-                }
-                else
-                {
-                    _ = Windows.System.Launcher.LaunchUriAsync(new Uri(url));
-                }
-            };
-
-            var updateRow = new StackPanel { Orientation = Orientation.Horizontal, Spacing = 8 };
-            updateRow.Children.Add(copyBtn);
-            updateRow.Children.Add(checkBtn);
-
-            var header = new StackPanel { Spacing = 0, Margin = new Thickness(0, 0, 0, 4) };
-            header.Children.Add(titleRow);
-            header.Children.Add(updateRow);
-            header.Children.Add(statusText);
-            header.Children.Add(updateBtn);
-
-            // ── Section helper ────────────────────────────────────────────────
-            static StackPanel MakeSection(string title)
-            {
-                var s = new StackPanel { Spacing = 4, Margin = new Thickness(0, 12, 0, 0) };
-                s.Children.Add(new NavigationViewItemSeparator { Margin = new Thickness(0, 0, 0, 8) });
-                s.Children.Add(new TextBlock
-                {
-                    Text       = title,
-                    FontSize   = 12,
-                    FontWeight = Microsoft.UI.Text.FontWeights.SemiBold,
-                    Opacity    = 0.8,
-                    Margin     = new Thickness(0, 0, 0, 4),
-                });
-                return s;
-            }
-
-            HyperlinkButton MakeLink(string text, string url) => new HyperlinkButton
-            {
-                NavigateUri = new Uri(url),
-                Padding     = new Thickness(0),
-                Content     = new StackPanel
-                {
-                    Orientation       = Orientation.Horizontal,
-                    Spacing           = 4,
-                    VerticalAlignment = VerticalAlignment.Center,
-                    Children          =
-                    {
-                        new TextBlock { Text = text, VerticalAlignment = VerticalAlignment.Center },
-                        new FontIcon
-                        {
-                            Glyph             = "",
-                            FontFamily        = new FontFamily("Segoe Fluent Icons"),
-                            FontSize          = 10,
-                            Opacity           = 0.6,
-                            VerticalAlignment = VerticalAlignment.Center,
-                        },
-                    }
-                },
-            };
-
-            var linksSection = MakeSection(R.Get("About_Links"));
-            linksSection.Children.Add(MakeLink(R.Get("About_Repository"), repoUrl));
-            linksSection.Children.Add(MakeLink(R.Get("Button_ReportIssue"), issueUrl));
-
-            var acksSection = MakeSection(R.Get("About_Acknowledgements"));
-            acksSection.Children.Add(MakeLink("TwitterTimelineLoader",
-                "https://chromewebstore.google.com/detail/twittertimelineloader/ipmgjpmedafkmmadinmeoannpofakpbh"));
-
-            var licenseSection = MakeSection(R.Get("About_License"));
-            licenseSection.Children.Add(new TextBlock { Text = "MIT License", IsTextSelectionEnabled = true, FontSize = 13 });
-
-            var componentsSection = MakeSection(R.Get("About_Components"));
-            componentsSection.Children.Add(new TextBlock
-            {
-                Text                   = $"{edgeChannel}  {edgeVersion}",
-                IsTextSelectionEnabled = true,
-                FontSize               = 13,
-                Opacity                = 0.8,
-            });
-
-            var panel = new StackPanel { MinWidth = 320 };
-            panel.Children.Add(header);
-            panel.Children.Add(linksSection);
-            panel.Children.Add(acksSection);
-            panel.Children.Add(licenseSection);
-            panel.Children.Add(componentsSection);
-
-            dlg = new ContentDialog
-            {
-                Title           = R.Get("About_Title"),
-                Content         = panel,
-                CloseButtonText = R.Get("Button_Close"),
-                XamlRoot        = Content.XamlRoot,
-                RequestedTheme  = ((FrameworkElement)Content).ActualTheme,
-            };
-            await ShowDialogAsync(dlg);
-        }
-
-                // ── Theme ─────────────────────────────────────────────────────────────
-
-        private async void PostBtn_Click(object _, RoutedEventArgs __)
-        {
-            if (_appSettings.OpenComposerInBrowser)
-                _ = Windows.System.Launcher.LaunchUriAsync(new Uri("https://x.com/compose/post"));
-            else
-                await OpenPostDialogAsync();
-        }
-
-        private async Task OpenPostDialogAsync(WebView2? senderWebView = null)
-        {
-            var webView = new WebView2 { Width = 500, MinHeight = 520 };
-
-            var dlg = new ContentDialog
-            {
-                Content         = webView,
-                CloseButtonText = R.Get("Button_Close"),
-                XamlRoot        = Content.XamlRoot,
-                RequestedTheme  = ((FrameworkElement)Content).ActualTheme
-            };
-
-            var env = _appSettings.SeparateComposeEnv
-                ? await GetOrCreateComposeEnvAsync()
-                : await GetOrCreateProfileEnvAsync("default");
-            await webView.EnsureCoreWebView2Async(env);
-
-            // テーマを適用
-            var root = (FrameworkElement)Content;
-            var scheme = root.ActualTheme switch
-            {
-                ElementTheme.Light => CoreWebView2PreferredColorScheme.Light,
-                ElementTheme.Dark  => CoreWebView2PreferredColorScheme.Dark,
-                _                  => CoreWebView2PreferredColorScheme.Auto,
-            };
-            webView.CoreWebView2.Profile.PreferredColorScheme = scheme;
-
-            bool composerReady = false;
-
-            webView.CoreWebView2.NavigationCompleted += async (s, args) =>
-            {
-                if (!args.IsSuccess) return;
-                composerReady = true;
-                await webView.CoreWebView2.ExecuteScriptAsync("""
-                    (function() {
-                        var id = 'xtv-compose-style';
-                        if (document.getElementById(id)) return;
-                        var s = document.createElement('style');
-                        s.id = id;
-                        s.textContent =
-                            '[data-testid="primaryColumn"],' +
-                            '[data-testid="sidebarColumn"],' +
-                            'header[role="banner"],' +
-                            '[data-testid="modalBackdrop"]' +
-                            '{display:none!important}';
-                        document.head.appendChild(s);
-                    })();
-                    """);
-            };
-
-            webView.CoreWebView2.NavigationStarting += (s, args) =>
-            {
-                if (composerReady && !args.Uri.Contains("/compose/post"))
-                    dlg.Hide();
-            };
-
-            webView.Source = new Uri("https://x.com/compose/post");
-
-            // WebView2 の Win32 HWND は XAML Popup より常に前面に描画されるため、
-            // ダイアログ表示中はタイムライン WebView2 を非表示にして Z-order 問題を回避する
-            foreach (var wv in _webViews)
-                wv.Visibility = Visibility.Collapsed;
-            try
-            {
-                await ShowDialogAsync(dlg);
-            }
-            finally
-            {
-                foreach (var wv in _webViews)
-                    wv.Visibility = Visibility.Visible;
-
-                // ダイアログを閉じた後、キーボードフォーカスを WebView2 に戻す
-                var target = senderWebView ?? _webViews.FirstOrDefault();
-                if (target is not null &&
-                    _webViewToPane.TryGetValue(target, out var pane) &&
-                    _paneToSetFocus.TryGetValue(pane, out var setFocus))
-                {
-                    setFocus();
-                }
-            }
-        }
-
-        // ── Keyboard shortcuts ────────────────────────────────────────────────
-
-        private void OnWebViewMessageReceived(WebView2 senderWebView, string message)
-        {
-            if (message.StartsWith("openTimestamp:") &&
-                Uri.TryCreate(message[14..], UriKind.Absolute, out var timestampUri))
-            {
-                _ = Windows.System.Launcher.LaunchUriAsync(timestampUri);
-                return;
-            }
-
-            switch (message)
-            {
-                case "focusNext": FocusAdjacentTimeline(senderWebView, +1); break;
-                case "focusPrev": FocusAdjacentTimeline(senderWebView, -1); break;
-                case "newPost":
-                    if (_appSettings.OpenComposerInBrowser)
-                        _ = Windows.System.Launcher.LaunchUriAsync(new Uri("https://x.com/compose/post"));
-                    else
-                        _ = OpenPostDialogAsync(senderWebView);
-                    break;
-            }
-        }
-
-        private void FocusAdjacentTimeline(WebView2 senderWebView, int direction)
-        {
-            if (!_webViewToPane.TryGetValue(senderWebView, out var senderPane)) return;
-            int idx  = TimelinePanel.Children.IndexOf(senderPane);
-            int next = idx + direction;
-            if (next < 0 || next >= TimelinePanel.Children.Count) return;
-            var targetPane = (Grid)TimelinePanel.Children[next];
-            if (_paneToSetFocus.TryGetValue(targetPane, out var setFocus))
-            {
-                setFocus();
-                targetPane.StartBringIntoView();
-            }
-        }
-
-        private void ApplyThemeToWebViews()
-        {
-            var root   = (FrameworkElement)Content;
-            var scheme = root.RequestedTheme switch
-            {
-                ElementTheme.Light => CoreWebView2PreferredColorScheme.Light,
-                ElementTheme.Dark  => CoreWebView2PreferredColorScheme.Dark,
-                _                  => CoreWebView2PreferredColorScheme.Auto,
-            };
-            foreach (var wv in _webViews)
-                if (wv.CoreWebView2 is not null)
-                    wv.CoreWebView2.Profile.PreferredColorScheme = scheme;
-        }
-
-        // ── Persistence ───────────────────────────────────────────────────────
-
-        private async Task SaveTimelinesAsync()
-        {
-            Directory.CreateDirectory(Path.GetDirectoryName(SaveFilePath)!);
-            var json = JsonSerializer.Serialize(_configs, JsonOptions);
-            await File.WriteAllTextAsync(SaveFilePath, json);
-        }
-
-        private async Task RestoreTimelinesAsync()
-        {
-            try
-            {
-                var json    = await File.ReadAllTextAsync(SaveFilePath);
-                var configs = JsonSerializer.Deserialize<List<TimelineConfig>>(json);
-                if (configs is not null)
-                    foreach (var cfg in configs)
-                        AddTimeline(cfg);
-            }
-            catch { /* ファイルが存在しない場合などは無視 */ }
-        }
-
-        // ── Drag & Drop ───────────────────────────────────────────────────────
-
-        private void MainArea_DragOver(object sender, DragEventArgs e)
-        {
-            if (e.DataView.Contains(StandardDataFormats.StorageItems))
-            {
-                e.AcceptedOperation          = DataPackageOperation.Link;
-                e.DragUIOverride.Caption     = R.Get("DragCaption");
-                e.DragUIOverride.IsGlyphVisible = true;
-            }
-            else
-            {
-                e.AcceptedOperation = DataPackageOperation.None;
-            }
-        }
-
-        private async void MainArea_Drop(object sender, DragEventArgs e)
-        {
-            if (!e.DataView.Contains(StandardDataFormats.StorageItems)) return;
-            var deferral = e.GetDeferral();
-            try
-            {
-                var items = await e.DataView.GetStorageItemsAsync();
-                foreach (var item in items)
-                {
-                    if (item is StorageFile file &&
-                        file.FileType.Equals(".url", StringComparison.OrdinalIgnoreCase))
-                    {
-                        var url = await ParseUrlShortcutAsync(file);
-                        if (url is not null && IsXUrl(url))
-                            AddTimeline(new TimelineConfig { Url = url });
-                    }
-                }
-            }
-            finally { deferral.Complete(); }
-        }
-
-        private static async Task<string?> ParseUrlShortcutAsync(StorageFile file)
-        {
-            try
-            {
-                var lines = await FileIO.ReadLinesAsync(file);
-                foreach (var line in lines)
-                    if (line.StartsWith("URL=", StringComparison.OrdinalIgnoreCase))
-                        return line[4..].Trim();
-            }
-            catch { }
-            return null;
-        }
-
-        private static bool IsXUrl(string url) =>
-            url.Contains("x.com",       StringComparison.OrdinalIgnoreCase) ||
-            url.Contains("twitter.com", StringComparison.OrdinalIgnoreCase);
-
-        private static readonly Color[] ProfileBadgeColors =
-        [
-            Color.FromArgb(255,  56, 142, 60),   // green
-            Color.FromArgb(255, 211,  47,  47),  // red
-            Color.FromArgb(255,  25, 118, 210),  // blue
-            Color.FromArgb(255, 156,  39, 176),  // purple
-            Color.FromArgb(255, 245, 124,   0),  // orange
-            Color.FromArgb(255,   0, 151, 167),  // teal
-            Color.FromArgb(255, 121,  85,  72),  // brown
-            Color.FromArgb(255,  63,  81, 181),  // indigo
-        ];
-
-        private static Color GetProfileColor(string profileId)
-        {
-            var hash = Math.Abs(profileId.GetHashCode());
-            return ProfileBadgeColors[hash % ProfileBadgeColors.Length];
-        }
-
-        private Color GetProfileColor(ProfileConfig? profile, string profileId)
-        {
-            if (profile?.BadgeColorIndex is int idx && idx >= 0 && idx < ProfileBadgeColors.Length)
-                return ProfileBadgeColors[idx];
-            return GetProfileColor(profileId);
-        }
-
-        private Border CreateProfileBadge(string profileId)
-        {
-            var showBadge = _profiles.Count > 1 && profileId != "default";
-            var profile = _profiles.FirstOrDefault(p => p.Id == profileId);
-            var name = profile?.Name ?? profileId;
-            var badgeText = profile?.BadgeText is { Length: > 0 } custom
-                ? custom
-                : (name.Length > 3 ? name[..3] : name);
-            var color = GetProfileColor(profile, profileId);
-
-            return new Border
-            {
-                Background    = new SolidColorBrush(color),
-                CornerRadius  = new CornerRadius(4),
-                Padding       = new Thickness(4, 1, 4, 1),
-                Margin        = new Thickness(0, 0, 6, 0),
-                VerticalAlignment = VerticalAlignment.Center,
-                Visibility    = showBadge ? Visibility.Visible : Visibility.Collapsed,
-                Child = new TextBlock
-                {
-                    Text       = badgeText,
-                    FontSize   = 10,
-                    FontWeight = Microsoft.UI.Text.FontWeights.SemiBold,
-                    Foreground = new SolidColorBrush(Microsoft.UI.Colors.White),
-                }
-            };
-        }
-
-        // ── AddTimeline ───────────────────────────────────────────────────────
-
-        private void AddTimeline(TimelineConfig cfg)
-        {
-            _configs.Add(cfg);
-            _ = SaveTimelinesAsync();
-
-            DropHintBorder.Visibility = Visibility.Collapsed;
-            TimelineScroll.Visibility = Visibility.Visible;
-
-            // Pane
-            var pane = new Grid
-            {
-                Width             = cfg.Width,
-                Margin            = new Thickness(4),
-                VerticalAlignment = VerticalAlignment.Stretch,
-                BorderThickness   = new Thickness(1),
-                CornerRadius      = new CornerRadius(8)
-            };
-            pane.RowDefinitions.Add(new RowDefinition { Height = GridLength.Auto });
-            pane.RowDefinitions.Add(new RowDefinition { Height = new GridLength(1, GridUnitType.Star) });
-
-            // Header
-            var headerGrid = new Grid { Padding = new Thickness(8, 4, 4, 4) };
-            headerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto }); // typeIcon
-            headerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto }); // profileBadge
-            headerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) }); // urlLabel
-            headerGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto }); // buttons
-            Grid.SetRow(headerGrid, 0);
-
-            // Theme
-            void ApplyPaneTheme(ElementTheme theme)
-            {
-                // Application.Current.Resources はアプリレベルのテーマを参照するため、
-                // 要素単位で RequestedTheme を設定している場合に正しい辞書が返らない。
-                // pane.ActualTheme（解決済み）を使い ThemeDictionaries を直接引く。
-                bool focused   = _focusedHeaderGrid == headerGrid;
-                var themeKey   = theme == ElementTheme.Light ? "Light" : "Default";
-                var themeDict  = (ResourceDictionary)Application.Current.Resources.ThemeDictionaries[themeKey];
-                pane.Background       = (Brush)themeDict["TimelinePaneBackgroundBrush"];
-                pane.BorderBrush      = (Brush)themeDict["TimelinePaneBorderBrush"];
-                headerGrid.Background = (Brush)themeDict[focused
-                    ? "TimelineHeaderFocusedBackgroundBrush"
-                    : "TimelineHeaderBackgroundBrush"];
-            }
-            ApplyPaneTheme(((FrameworkElement)Content).ActualTheme);
-            pane.ActualThemeChanged += (s, _) => ApplyPaneTheme(pane.ActualTheme);
-
-            // URL label
-            string displayText = cfg.Url;
-            if (Uri.TryCreate(cfg.Url, UriKind.Absolute, out var uri))
-                displayText = uri.Host + uri.PathAndQuery;
-
-            var typeIcon = new FontIcon
-            {
-                Glyph             = GetTimelineGlyph(cfg.Url),
-                FontFamily        = new FontFamily("Segoe Fluent Icons"),
-                FontSize          = 14,
-                Opacity           = 0.8,
-                Margin            = new Thickness(0, 0, 8, 0),
-                VerticalAlignment = VerticalAlignment.Center
-            };
-            Grid.SetColumn(typeIcon, 0);
-            var hardReloadTooltip = new ToolTip();
-            ToolTipService.SetToolTip(typeIcon, hardReloadTooltip);
-
-            // Profile badge
-            var profileBadge = CreateProfileBadge(cfg.ProfileId);
-            Grid.SetColumn(profileBadge, 1);
-
-            var urlLabel = new TextBlock
-            {
-                Text              = displayText,
-                FontSize          = 12,
-                TextTrimming      = TextTrimming.CharacterEllipsis,
-                VerticalAlignment = VerticalAlignment.Center,
-                Opacity           = 0.8
-            };
-            Grid.SetColumn(urlLabel, 2);
-
-            // WebView2
-            var webView = new WebView2
-            {
-                VerticalAlignment   = VerticalAlignment.Stretch,
-                HorizontalAlignment = HorizontalAlignment.Stretch
-            };
-            Grid.SetRow(webView, 1);
-
-            // Buttons
-            var buttonPanel = new StackPanel
-            {
-                Orientation       = Orientation.Horizontal,
-                Spacing           = 4,
-                VerticalAlignment = VerticalAlignment.Center
-            };
-            Grid.SetColumn(buttonPanel, 3);
-
-            var settingsBtn = new Button
-            {
-                Content = new FontIcon { Glyph = "\uE713", FontFamily = new FontFamily("Segoe Fluent Icons"), FontSize = 14 },
-                Width = 28, Height = 26,
-                Padding = new Thickness(0)
-            };
-            ToolTipService.SetToolTip(settingsBtn, R.Get("Pane_Settings_Tooltip"));
-            AutomationProperties.SetName(settingsBtn, R.Get("Pane_Settings_Tooltip"));
-            AutomationProperties.SetName(webView, displayText);
-
-            buttonPanel.Children.Add(settingsBtn);
-
-            headerGrid.Children.Add(typeIcon);
-            headerGrid.Children.Add(profileBadge);
-            headerGrid.Children.Add(urlLabel);
-            headerGrid.Children.Add(buttonPanel);
-
-            pane.Children.Add(headerGrid);
-            pane.Children.Add(webView);
-            TimelinePanel.Children.Add(pane);
-            _webViews.Add(webView);
-            _webViewToPane[webView] = pane;
-
-            // ── Focus ─────────────────────────────────────────────────────────
-
-            Action refreshHeader = () => ApplyPaneTheme(pane.ActualTheme);
-            _headerRefreshers.Add(refreshHeader);
-
-            void SetFocus()
-            {
-                if (_focusedHeaderGrid is not null && _homeHeaderGrids.Contains(_focusedHeaderGrid) && !_homeHeaderGrids.Contains(headerGrid))
-                    RestartAutoActivateTimer();  // ホームから別タイムラインへ
-                _focusedHeaderGrid = headerGrid;
-                foreach (var r in _headerRefreshers) r();
-                webView.Focus(FocusState.Programmatic);
-            }
-            _paneToSetFocus[pane] = SetFocus;
-
-            headerGrid.Tapped        += (s, e) => SetFocus();
-            headerGrid.DoubleTapped  += (s, e) =>
-            {
-                SetFocus();
-                webView.Source = new Uri(cfg.Url);
-            };
-            webView.GotFocus   += (s, e) =>
-            {
-                if (_focusedHeaderGrid is not null && _homeHeaderGrids.Contains(_focusedHeaderGrid) && !_homeHeaderGrids.Contains(headerGrid))
-                    RestartAutoActivateTimer();  // ホームから別タイムラインへ
-                _focusedHeaderGrid = headerGrid;
-                foreach (var r in _headerRefreshers) r();
-            };
-            webView.PointerEntered += (s, e) => { _pointerOverWebViews.Add(webView);    EvaluateHardReloadPause(webView); };
-            webView.PointerExited  += (s, e) =>
-            {
-                _pointerOverWebViews.Remove(webView);
-                EvaluateHardReloadPause(webView);
-                if (_pointerOverWebViews.Count == 0) RestartAutoActivateTimer();
-            };
-
-            bool isHomeTimeline = Uri.TryCreate(cfg.Url, UriKind.Absolute, out var cfgUri)
-                               && cfgUri.AbsolutePath.StartsWith("/home", StringComparison.OrdinalIgnoreCase);
-            if (isHomeTimeline) _homeHeaderGrids.Add(headerGrid);
-
-            _hardReloadUiUpdaters[webView] = () => UpdateHardReloadTooltip(webView, hardReloadTooltip);
-            EnsureHardReloadUiTimer();
-
-            // ── Drag & Drop reorder ───────────────────────────────────────────
-
-            headerGrid.CanDrag = true;
-            headerGrid.DragStarting += (s, args) =>
-            {
-                _draggingPane = pane;
-                args.Data.SetText("xtv-pane");
-            };
-
-            pane.AllowDrop = true;
-            pane.DragOver  += (s, args) =>
-            {
-                if (_draggingPane is not null && _draggingPane != pane)
-                {
-                    args.AcceptedOperation = DataPackageOperation.Move;
-                    args.Handled = true;
-                }
-            };
-            pane.Drop += (s, args) =>
-            {
-                if (_draggingPane is null || _draggingPane == pane) return;
-                args.Handled = true;
-
-                var dragging = _draggingPane;
-
-                int from = TimelinePanel.Children.IndexOf(dragging);
-                int to   = TimelinePanel.Children.IndexOf(pane);
-                if (from < 0 || to < 0) return;
-
-                TimelinePanel.Children.RemoveAt(from);
-                TimelinePanel.Children.Insert(to, dragging);
-
-                var cfg2 = _configs[from];
-                _configs.RemoveAt(from);
-                _configs.Insert(to, cfg2);
-
-                _ = SaveTimelinesAsync();
-                dragging.Opacity = 1.0;
-                _draggingPane = null;
-
-                // 視覚ツリーへの再挿入後、WebView2 の Win32 HWND を再アンカーさせる
-                dragging.Visibility = Visibility.Collapsed;
-                dragging.UpdateLayout();
-                dragging.Visibility = Visibility.Visible;
-            };
-            pane.DragLeave += (s, args) => pane.Opacity = 1.0;
-            headerGrid.DragStarting += (s, args) => pane.Opacity = 0.5;
-
-            // ── Settings dialog ───────────────────────────────────────────────
-
-            settingsBtn.Click += async (s, e2) =>
-            {
-                var widthBox = new NumberBox
-                {
-                    Value                   = cfg.Width,
-                    Minimum                 = 100,
-                    Maximum                 = 2000,
-                    SmallChange             = 10,
-                    LargeChange             = 50,
-                    SpinButtonPlacementMode = NumberBoxSpinButtonPlacementMode.Inline,
-                    Width                   = 160
-                };
-
-                var hideSidebarToggle = new ToggleSwitch
-                {
-                    IsOn       = cfg.HideSidebar,
-                    OnContent  = R.Get("Toggle_Hide"),
-                    OffContent = R.Get("Toggle_Show")
-                };
-
-                var hideComposeToggle = new ToggleSwitch
-                {
-                    IsOn       = cfg.HideCompose,
-                    OnContent  = R.Get("Toggle_Hide"),
-                    OffContent = R.Get("Toggle_Show")
-                };
-
-                var listHeaderApplicable = IsListHeaderApplicable(cfg.Url);
-                var hideListHeaderToggle = new ToggleSwitch
-                {
-                    IsOn       = cfg.HideListHeader,
-                    IsEnabled  = listHeaderApplicable,
-                    OnContent  = R.Get("Toggle_Hide"),
-                    OffContent = R.Get("Toggle_Show")
-                };
-                var hideListHeaderLabel = new TextBlock
-                {
-                    Text    = R.Get("Timeline_ListHeader"),
-                    Margin  = new Thickness(0, 8, 0, 0),
-                    Opacity = listHeaderApplicable ? 1.0 : 0.4
-                };
-
-                var hardReloadToggle = new ToggleSwitch
-                {
-                    IsOn       = cfg.HardReloadEnabled,
-                    OnContent  = R.Get("Toggle_On"),
-                    OffContent = R.Get("Toggle_Off"),
-                };
-                var hardReloadIntervalBox = new NumberBox
-                {
-                    Value                   = cfg.HardReloadInterval,
-                    Minimum                 = 1,
-                    Maximum                 = 60,
-                    SpinButtonPlacementMode = NumberBoxSpinButtonPlacementMode.Inline,
-                    Width                   = 160,
-                    IsEnabled               = cfg.HardReloadEnabled,
-                };
-                hardReloadToggle.Toggled += (_, _) =>
-                    hardReloadIntervalBox.IsEnabled = hardReloadToggle.IsOn;
-                var reloadLabel = new TextBlock
-                {
-                    Text   = R.Get("Timeline_ReloadInterval"),
-                    Margin = new Thickness(0, 8, 0, 0),
-                };
-
-                var deleteBtn = new Button
-                {
-                    Content             = R.Get("Pane_Delete"),
-                    HorizontalAlignment = HorizontalAlignment.Stretch,
-                    Margin              = new Thickness(0, 16, 0, 0),
-                    Foreground          = (Brush)Application.Current.Resources["SystemFillColorCriticalBrush"],
-                };
-
-                var profileBox = new ComboBox { MinWidth = 200 };
-                foreach (var p in _profiles)
-                    profileBox.Items.Add(new ComboBoxItem { Content = p.Name, Tag = p.Id });
-                profileBox.SelectedItem = profileBox.Items
-                    .OfType<ComboBoxItem>()
-                    .FirstOrDefault(i => (string)i.Tag == cfg.ProfileId)
-                    ?? profileBox.Items.OfType<ComboBoxItem>().FirstOrDefault();
-
-                var panel = new StackPanel { Spacing = 8 };
-                panel.Children.Add(new TextBlock { Text = R.Get("Timeline_Profile") });
-                panel.Children.Add(profileBox);
-                panel.Children.Add(new TextBlock { Text = R.Get("Timeline_Width"), Margin = new Thickness(0, 8, 0, 0) });
-                panel.Children.Add(widthBox);
-                panel.Children.Add(new TextBlock
-                {
-                    Text   = R.Get("Timeline_Sidebar"),
-                    Margin = new Thickness(0, 8, 0, 0)
-                });
-                panel.Children.Add(hideSidebarToggle);
-                panel.Children.Add(new TextBlock
-                {
-                    Text   = R.Get("Timeline_Compose"),
-                    Margin = new Thickness(0, 8, 0, 0)
-                });
-                panel.Children.Add(hideComposeToggle);
-                panel.Children.Add(hideListHeaderLabel);
-                panel.Children.Add(hideListHeaderToggle);
-                panel.Children.Add(reloadLabel);
-                panel.Children.Add(hardReloadToggle);
-                panel.Children.Add(hardReloadIntervalBox);
-                panel.Children.Add(new NavigationViewItemSeparator { Margin = new Thickness(0, 8, 0, 0) });
-                panel.Children.Add(deleteBtn);
-
-                var dlg = new ContentDialog
-                {
-                    Title             = R.Get("Timeline_Settings_Title"),
-                    Content           = panel,
-                    PrimaryButtonText = R.Get("Button_Apply"),
-                    CloseButtonText   = R.Get("Button_Cancel"),
-                    DefaultButton     = ContentDialogButton.Primary,
-                    XamlRoot          = Content.XamlRoot
-                };
-
-                bool shouldDelete = false;
-                deleteBtn.Click += (_, _) => { shouldDelete = true; dlg.Hide(); };
-
-                var result = await ShowDialogAsync(dlg);
-
-                if (shouldDelete)
-                {
-                    CleanupWebView(webView);
-                    if (_hardReloadUiUpdaters.Count == 0)
-                    {
-                        _hardReloadUiTimer?.Stop();
-                        _hardReloadUiTimer = null;
-                    }
-                    _configs.Remove(cfg);
-                    _paneToSetFocus.Remove(pane);
-                    _headerRefreshers.Remove(refreshHeader);
-                    if (_focusedHeaderGrid == headerGrid)
-                    {
-                        _focusedHeaderGrid = null;
-                        foreach (var r in _headerRefreshers) r();
-                    }
-                    await SaveTimelinesAsync();
-
-                    TimelinePanel.Children.Remove(pane);
-                    if (TimelinePanel.Children.Count == 0)
-                    {
-                        TimelineScroll.Visibility = Visibility.Collapsed;
-                        DropHintBorder.Visibility = Visibility.Visible;
-                    }
-                }
-                else if (result == ContentDialogResult.Primary)
-                {
-                    var prevProfileId = cfg.ProfileId;
-                    cfg.ProfileId = (profileBox.SelectedItem as ComboBoxItem)?.Tag as string ?? "default";
-
-                    cfg.Width  = Math.Clamp(widthBox.Value, 100, 2000);
-                    pane.Width = cfg.Width;
-
-                    cfg.HideSidebar = hideSidebarToggle.IsOn;
-                    cfg.HideCompose = hideComposeToggle.IsOn;
-                    cfg.HideListHeader = hideListHeaderToggle.IsOn;
-                    cfg.HardReloadEnabled  = hardReloadToggle.IsOn;
-                    cfg.HardReloadInterval = (int)Math.Clamp(hardReloadIntervalBox.Value, 1, 60);
-
-                    if (prevProfileId != cfg.ProfileId)
-                    {
-                        CleanupWebView(webView);
-                        pane.Children.Remove(webView);
-
-                        webView = new WebView2
-                        {
-                            VerticalAlignment   = VerticalAlignment.Stretch,
-                            HorizontalAlignment = HorizontalAlignment.Stretch
-                        };
-                        Grid.SetRow(webView, 1);
-                        pane.Children.Add(webView);
-                        _webViews.Add(webView);
-                        _webViewToPane[webView] = pane;
-
-                        var newBadge = CreateProfileBadge(cfg.ProfileId);
-                        Grid.SetColumn(newBadge, 1);
-                        headerGrid.Children.Remove(profileBadge);
-                        headerGrid.Children.Add(newBadge);
-                        profileBadge = newBadge;
-
-                        Debug.WriteLine($"[Profile] WebView2 recreated for profile switch: {prevProfileId} -> {cfg.ProfileId}");
-                        _ = InitWebViewAsync(webView, cfg);
-                    }
-                    else
-                    {
-                        if (webView.CoreWebView2 is not null)
-                        {
-                            await ApplyHideSidebarAsync(webView, cfg.HideSidebar);
-                            await ApplyHideComposeAsync(webView, cfg.HideCompose);
-                            await ApplyHideListHeaderAsync(webView, cfg.HideListHeader);
-                        }
-                    }
-
-                    StartHardReloadTimer(webView, cfg);
-                    await SaveTimelinesAsync();
-                }
-            };
-
-            _ = InitWebViewAsync(webView, cfg);
-        }
-
-        private void StartHardReloadTimer(WebView2 wv, TimelineConfig cfg)
-        {
-            StopHardReloadTimer(wv);
-            if (!cfg.HardReloadEnabled) return;
-
-            var timer = new DispatcherTimer { Interval = TimeSpan.FromMinutes(cfg.HardReloadInterval) };
-            timer.Tick += (_, _) =>
-            {
-                wv.CoreWebView2?.Reload();
-                _hardReloadStartTimes[wv] = DateTimeOffset.Now;
-            };
-            _hardReloadStartTimes[wv] = DateTimeOffset.Now;
-            timer.Start();
-            _hardReloadTimers[wv] = timer;
-        }
-
-        private void StopHardReloadTimer(WebView2 wv)
-        {
-            if (_hardReloadTimers.Remove(wv, out var t)) t.Stop();
-            _hardReloadStartTimes.Remove(wv);
-        }
-
-        // WebView2 インスタンスに紐づくすべてのリソースを解放し、CoreWebView2 を閉じる。
-        // タイムライン削除・プロファイル切り替え・ウィンドウクローズ時に必ず呼ぶこと。
-        private void CleanupWebView(WebView2 wv)
-        {
-            var source = wv.CoreWebView2?.Source ?? "(not initialized)";
-            Debug.WriteLine($"[WebView2] CleanupWebView: source={source}");
-
-            StopHardReloadTimer(wv);
-            _hardReloadUiUpdaters.Remove(wv);
-            _pointerOverWebViews.Remove(wv);
-            _urlDivergedWebViews.Remove(wv);
-            _webViews.Remove(wv);
-            _webViewToPane.Remove(wv);
-            try
-            {
-                wv.Close();
-                Debug.WriteLine($"[WebView2] Closed: source={source}");
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"[WebView2] Close failed: source={source}, error={ex.Message}");
-            }
-        }
-
-        private void EvaluateHardReloadPause(WebView2 wv)
-        {
-            if (!_hardReloadTimers.TryGetValue(wv, out var t)) return;
-            bool shouldPause = _pointerOverWebViews.Contains(wv) || _urlDivergedWebViews.Contains(wv);
-            if (shouldPause && t.IsEnabled)
-                t.Stop();
-            else if (!shouldPause && !t.IsEnabled)
-            {
-                _hardReloadStartTimes[wv] = DateTimeOffset.Now;
-                t.Start();
-            }
-        }
-
-        private string GetAutoActivateTooltipText(WebView2 wv)
-        {
-            if (_autoActivateTimer is null)
-                return R.Get("AutoActivate_Disabled");
-            if (_pointerOverWebViews.Count > 0 || _dialogOpenCount > 0)
-                return R.Get("AutoActivate_Paused");
-            if (_urlDivergedWebViews.Contains(wv))
-                return R.Get("AutoActivate_Paused_Nav");
-            var remaining = _autoActivateTimer.Interval - (DateTimeOffset.Now - _autoActivateStartTime);
-            return remaining > TimeSpan.Zero
-                ? string.Format(R.Get("AutoActivate_Active"), (int)remaining.TotalMinutes, remaining.Seconds.ToString("D2"))
-                : string.Empty;
-        }
-
         private async Task<ContentDialogResult> ShowDialogAsync(ContentDialog dlg)
         {
             _dialogOpenCount++;
@@ -1753,608 +325,6 @@ namespace XTimelineViewer
             if (!Uri.TryCreate(baseUrl,    UriKind.Absolute, out var @base)) return false;
             return string.Equals(cur.Host,         @base.Host,         StringComparison.OrdinalIgnoreCase)
                 && string.Equals(cur.AbsolutePath, @base.AbsolutePath, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private string GetHardReloadTooltipText(WebView2 wv)
-        {
-            if (!_hardReloadTimers.TryGetValue(wv, out var t))
-                return R.Get("HardReload_Disabled");
-            if (!t.IsEnabled)
-                return _urlDivergedWebViews.Contains(wv)
-                    ? R.Get("HardReload_Paused_Nav")
-                    : R.Get("HardReload_Paused");
-            if (_hardReloadStartTimes.TryGetValue(wv, out var start))
-            {
-                var remaining = t.Interval - (DateTimeOffset.Now - start);
-                if (remaining > TimeSpan.Zero)
-                    return string.Format(R.Get("HardReload_Active"), (int)remaining.TotalMinutes, remaining.Seconds.ToString("D2"));
-            }
-            return string.Empty;
-        }
-
-        private void UpdateHardReloadTooltip(WebView2 wv, ToolTip tooltip)
-        {
-            tooltip.Content = GetHardReloadTooltipText(wv) is { Length: > 0 } text ? text : null;
-        }
-
-        private void UpdateAutoActivateLabel()
-        {
-            if (!_appSettings.ShowAutoActivateLabel)
-            {
-                AutoActivateLabel.Visibility = Visibility.Collapsed;
-                return;
-            }
-
-            var homeWv = _webViews.FirstOrDefault(wv =>
-                wv.CoreWebView2 is not null &&
-                Uri.TryCreate(wv.CoreWebView2.Source, UriKind.Absolute, out var u) &&
-                u.AbsolutePath.StartsWith("/home", StringComparison.OrdinalIgnoreCase));
-
-            if (homeWv is null)
-            {
-                AutoActivateLabel.Visibility = Visibility.Collapsed;
-                return;
-            }
-
-            var state = GetAutoActivateTooltipText(homeWv);
-            if (string.IsNullOrEmpty(state))
-            {
-                AutoActivateLabel.Visibility = Visibility.Collapsed;
-                return;
-            }
-
-            AutoActivateLabel.Text       = string.Format(R.Get("AutoActivateLabel_Format"), state);
-            AutoActivateLabel.Visibility = Visibility.Visible;
-        }
-
-        private void EnsureHardReloadUiTimer()
-        {
-            if (_hardReloadUiTimer is not null) return;
-            _hardReloadUiTimer = new DispatcherTimer { Interval = TimeSpan.FromSeconds(1) };
-            _hardReloadUiTimer.Tick += (_, _) =>
-            {
-                foreach (var (wv, update) in _hardReloadUiUpdaters) update();
-                UpdateAutoActivateLabel();
-            };
-            _hardReloadUiTimer.Start();
-        }
-
-        // ── Update check ─────────────────────────────────────────────────────
-
-        private static bool IsUpdateAvailable(Version current, Version latest)
-        {
-            if (PackageContext.IsPackaged)
-                return latest.Major > current.Major
-                    || (latest.Major == current.Major && latest.Minor > current.Minor);
-            return latest > current;
-        }
-
-        private static async Task<(Version version, string tag, string releaseUrl)> FetchLatestReleaseAsync()
-        {
-            using var client = new HttpClient();
-            client.DefaultRequestHeaders.Add("User-Agent", "XTimelineViewer");
-            var json = await client.GetStringAsync(
-                "https://api.github.com/repos/daruyanagi/XTimelineViewer/releases/latest");
-            using var doc = JsonDocument.Parse(json);
-            var tag = doc.RootElement.GetProperty("tag_name").GetString()!;
-            var url = doc.RootElement.GetProperty("html_url").GetString()!;
-            return (new Version(tag.TrimStart('v')), tag, url);
-        }
-
-        private async Task CheckForUpdatesInBackgroundAsync()
-        {
-            await Task.Delay(3000);
-            if (!NetworkInterface.GetIsNetworkAvailable()) return;
-
-            if (_appSettings.LastUpdateCheck is { } raw
-                && DateTime.TryParse(raw, null, DateTimeStyles.RoundtripKind, out var last)
-                && (DateTime.UtcNow - last).TotalDays < 7)
-                return;
-
-            try
-            {
-                var (latest, tag, _) = await FetchLatestReleaseAsync();
-                var current = Assembly.GetExecutingAssembly().GetName().Version!;
-                _appSettings.LastUpdateCheck     = DateTime.UtcNow.ToString("O");
-                _appSettings.CachedLatestVersion = IsUpdateAvailable(current, latest) ? tag : null;
-                SaveSettings();
-                UpdateMenuUpdateBadge();
-            }
-            catch { }
-        }
-
-        private void UpdateMenuUpdateBadge()
-        {
-            UpdateBadgeDot.Visibility = _appSettings.CachedLatestVersion is not null
-                ? Visibility.Visible
-                : Visibility.Collapsed;
-        }
-
-        private static string? FindWinget()
-        {
-            var candidate = Path.Combine(
-                Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-                "Microsoft", "WindowsApps", "winget.exe");
-            if (File.Exists(candidate)) return candidate;
-
-            return Environment.GetEnvironmentVariable("PATH")
-                ?.Split(';', StringSplitOptions.RemoveEmptyEntries)
-                .Select(d => Path.Combine(d, "winget.exe"))
-                .FirstOrDefault(File.Exists);
-        }
-
-        private async Task SelfUpdateViaWingetAsync(string releaseUrl)
-        {
-            var confirmDlg = new ContentDialog
-            {
-                Title             = R.Get("CheckUpdate_WingetTitle"),
-                Content           = new TextBlock
-                {
-                    Text         = R.Get("CheckUpdate_WingetBody"),
-                    TextWrapping = TextWrapping.Wrap,
-                },
-                PrimaryButtonText = R.Get("CheckUpdate_WingetConfirm"),
-                CloseButtonText   = R.Get("Button_Cancel"),
-                XamlRoot          = Content.XamlRoot,
-                RequestedTheme    = ((FrameworkElement)Content).ActualTheme,
-            };
-
-            if (await ShowDialogAsync(confirmDlg) != ContentDialogResult.Primary) return;
-
-            var winget = FindWinget();
-            if (winget is null)
-            {
-                _ = Windows.System.Launcher.LaunchUriAsync(new Uri(releaseUrl));
-                return;
-            }
-
-            Process.Start(new ProcessStartInfo
-            {
-                FileName        = "cmd.exe",
-                Arguments       = "/c timeout /t 2 /nobreak > nul && winget upgrade daruyanagi.XTimelineViewer",
-                UseShellExecute = true,
-            });
-            Application.Current.Exit();
-        }
-
-        // ── WebView2 init ─────────────────────────────────────────────────────
-
-        private static string GetTimelineGlyph(string url)
-        {
-            if (!Uri.TryCreate(url, UriKind.Absolute, out var uri)) return "";
-            var p = uri.AbsolutePath;
-            if (p.StartsWith("/home"))                                return ""; // Home
-            if (p.StartsWith("/notifications"))                       return ""; // Bell
-            if (p.StartsWith("/search") || p.StartsWith("/explore")) return ""; // Search
-            if (p == "/bookmarks" || p.StartsWith("/bookmarks/") ||
-                p == "/i/bookmarks" || p.StartsWith("/i/bookmarks/")) return ""; // Bookmark
-            if (p.StartsWith("/i/lists/"))                            return ""; // BulletedList
-            if (p.StartsWith("/messages"))                            return ""; // Chat
-            if (System.Text.RegularExpressions.Regex.IsMatch(p, @"^/[^/]+$")) return ""; // Contact
-            return ""; // Globe
-        }
-
-        private static bool IsProfilePath(string p) =>
-            System.Text.RegularExpressions.Regex.IsMatch(p, @"^/[A-Za-z0-9_]+$") &&
-            !p.StartsWith("/home") && !p.StartsWith("/notifications") &&
-            !p.StartsWith("/search") && !p.StartsWith("/explore") &&
-            !p.StartsWith("/bookmarks") && !p.StartsWith("/messages") &&
-            !p.StartsWith("/i/");
-
-        private static bool IsListHeaderApplicable(string url)
-        {
-            if (!Uri.TryCreate(url, UriKind.Absolute, out var uri)) return false;
-            var p = uri.AbsolutePath;
-            return p.StartsWith("/notifications") ||
-                   p.StartsWith("/search")        ||
-                   p.StartsWith("/explore")       ||
-                   p == "/bookmarks" || p.StartsWith("/bookmarks/") ||
-                   p == "/i/bookmarks" || p.StartsWith("/i/bookmarks/") ||
-                   p.StartsWith("/i/lists/")      ||
-                   IsProfilePath(p);
-        }
-
-        private static string BuildHideListHeaderJs(bool hide) => $$"""
-            (function(hide){
-                var id='xtv-hide-list-header';
-                var s=document.getElementById(id);
-                if(hide){
-                    var path=window.location.pathname;
-                    var css=[];
-                    // 通知: 「通知」タイトル＋設定ギアを含むヘッダーブロックを非表示
-                    // 直接子セレクタで深さを固定し、全祖先にマッチしないよう限定する
-                    if(/^\/notifications/.test(path))
-                        css.push('div:has(>div>div>div>div>div>a[data-testid="settingsAppBar"]){display:none!important}');
-                    // 検索: 検索ボックス＋戻るボタンを含むヘッダーブロックを非表示
-                    if(/^\/(search|explore)/.test(path))
-                        css.push('div:has(>div>div>div>div>div>button[data-testid="app-bar-back"]){display:none!important}');
-                    // リスト: 上部ナビバー＋リスト情報カード（バナー・作成者・メンバー数）を非表示
-                    if(/^\/i\/lists\//.test(path)){
-                        css.push('div:has(>div>div>div>div>div>button[data-testid="app-bar-back"]){display:none!important}');
-                        css.push('[data-testid="cellInnerDiv"]:has(a[href*="/i/lists/"][href$="/members"]){display:none!important}');
-                    }
-                    // ブックマーク: 上部ナビバー＋nth-child(1)ブロック＋空フォルダの説明文を非表示
-                    if(/^\/(i\/bookmarks|bookmarks)/.test(path)){
-                        css.push('div:has(>div>div>div>div>div>button[data-testid="app-bar-back"]){display:none!important}');
-                        css.push('#react-root main section>div>div>div:nth-child(1){display:none!important}');
-                        css.push('[data-testid="emptyState"]{display:none!important}');
-                    }
-                    // プロフィール: ナビバー＋プロフィール情報カード（バナー・アバター・自己紹介・フォロー数）を非表示
-                    if(/^\/[A-Za-z0-9_]+$/.test(path) &&
-                       !/^\/(home|notifications|search|explore|bookmarks|messages|i)/.test(path)){
-                        css.push('div:has(>div>div>div>div>div>button[data-testid="app-bar-back"]){display:none!important}');
-                        css.push('div:has(>a[href$="/header_photo"]){display:none!important}');
-                    }
-                    if(css.length){
-                        if(!s){s=document.createElement('style');s.id=id;document.head.appendChild(s);}
-                        s.textContent=css.join('');
-                    }else{
-                        if(s)s.remove();
-                    }
-                }else{
-                    if(s)s.remove();
-                }
-            })({{(hide ? "true" : "false")}});
-            """;
-
-        private static async Task ApplyHideListHeaderAsync(
-            Microsoft.UI.Xaml.Controls.WebView2 webView, bool hide)
-        {
-            await webView.CoreWebView2.ExecuteScriptAsync(BuildHideListHeaderJs(hide));
-        }
-
-        private static string BuildHideSidebarJs(bool hide) => $$"""
-            (function(hide){
-                var id='xtv-hide-sidebar';
-                var s=document.getElementById(id);
-                if(hide){
-                    if(!s){s=document.createElement('style');s.id=id;
-                           s.textContent='header[role="banner"]{display:none!important}';
-                           document.head.appendChild(s);}
-                }else{
-                    if(s)s.remove();
-                }
-            })({{(hide ? "true" : "false")}});
-            """;
-
-        private static async Task ApplyHideSidebarAsync(
-            Microsoft.UI.Xaml.Controls.WebView2 webView, bool hide)
-        {
-            await webView.CoreWebView2.ExecuteScriptAsync(BuildHideSidebarJs(hide));
-        }
-
-        private static async Task ApplyAutoShowNewPostsAsync(WebView2 webView, string cfgUrl)
-        {
-            if (!Uri.TryCreate(cfgUrl, UriKind.Absolute, out var uri)) return;
-            if (!uri.AbsolutePath.TrimEnd('/').Equals("/home", StringComparison.OrdinalIgnoreCase)) return;
-
-            // 「（数字） 件のポストを表示」を含む button 要素を監視して自動でクリックするスクリプト
-            await webView.CoreWebView2.ExecuteScriptAsync("""
-                (function() {
-                    var observer = new MutationObserver(function(mutations) {
-                        mutations.forEach(function(mutation) {
-                            mutation.addedNodes.forEach(function(node) {
-                                if (node.nodeType === Node.ELEMENT_NODE) {
-                                    var btn = node.matches('button') ? node : node.querySelector('button');
-                                    if (btn && /件のポストを表示/.test(btn.textContent)) {
-                                        btn.click();
-                                    }
-                                }
-                            });
-                        });
-                    });
-                    observer.observe(document.body, { childList: true, subtree: true });
-                })();
-                """);
-        }
-
-        private static bool EffectiveHideCompose(TimelineConfig cfg, string currentUrl) =>
-            cfg.HideCompose && !currentUrl.Contains("compose/post", StringComparison.OrdinalIgnoreCase);
-
-        private static string BuildHideComposeJs(bool hide) => $$"""
-            (function(hide){
-                var id='xtv-hide-compose';
-                var s=document.getElementById(id);
-                if(hide){
-                    if(!s){s=document.createElement('style');s.id=id;
-                           s.textContent='.r-1h8ys4a{display:none!important}';
-                           document.head.appendChild(s);}
-                }else{
-                    if(s)s.remove();
-                }
-            })({{(hide ? "true" : "false")}});
-            """;
-
-        private static async Task ApplyHideComposeAsync(
-            Microsoft.UI.Xaml.Controls.WebView2 webView, bool hide)
-        {
-            await webView.CoreWebView2.ExecuteScriptAsync(BuildHideComposeJs(hide));
-        }
-
-        private async Task LoadExtensionsAsync(WebView2 webView)
-        {
-            if (_extensionsLoaded) return;
-            _extensionsLoaded = true;
-
-            // MSIX パッケージ内の extensions は WindowsApps 配下に置かれ WebView2 から直接アクセスできない。
-            // LocalState へコピーしてから読み込む。アンパッケージド環境は BaseDirectory を使う。
-            var extensionsDir = GetExtensionsDir();
-            if (!Directory.Exists(extensionsDir)) return;
-
-            var errors = new System.Text.StringBuilder();
-
-            foreach (var extDir in Directory.GetDirectories(extensionsDir))
-            {
-                try
-                {
-                    var ext = await webView.CoreWebView2.Profile.AddBrowserExtensionAsync(extDir);
-                    AddExtensionButton(ext, extDir);
-                }
-                catch (Exception ex)
-                {
-                    errors.AppendLine($"・{Path.GetFileName(extDir)}");
-                    errors.AppendLine($"  {ex}");
-                }
-            }
-
-            if (errors.Length > 0)
-            {
-                var dlg = new ContentDialog
-                {
-                    Title           = R.Get("ExtLoadError_Title"),
-                    Content         = new ScrollViewer
-                    {
-                        MaxHeight = 300,
-                        Content   = new TextBlock
-                        {
-                            Text       = errors.ToString().TrimEnd()
-                            + "\n\n" + webView.CoreWebView2.Environment.BrowserVersionString,
-                            FontFamily = new FontFamily("Cascadia Mono, Consolas, Courier New"),
-                            FontSize   = 12,
-                            IsTextSelectionEnabled = true,
-                            TextWrapping = TextWrapping.Wrap
-                        }
-                    },
-                    CloseButtonText = R.Get("Button_Close"),
-                    XamlRoot        = Content.XamlRoot
-                };
-                await ShowDialogAsync(dlg);
-            }
-        }
-
-        private void AddExtensionButton(CoreWebView2BrowserExtension ext, string extDir)
-        {
-            // マニフェストから名前、options_ui.page、アイコン、homepage_url を取得
-            string name          = ext.Name;
-            string? optPage      = null;
-            string? iconPath     = null;
-            string? homepageUrl  = null;
-            var manifestPath = Path.Combine(extDir, "manifest.json");
-            if (File.Exists(manifestPath))
-            {
-                using var doc = JsonDocument.Parse(File.ReadAllText(manifestPath));
-                var root = doc.RootElement;
-                if (root.TryGetProperty("options_ui", out var optUi) &&
-                    optUi.TryGetProperty("page", out var page))
-                    optPage = page.GetString();
-                if (root.TryGetProperty("icons", out var icons))
-                {
-                    foreach (var size in new[] { "16", "32", "48", "128" })
-                    {
-                        if (icons.TryGetProperty(size, out var iconProp))
-                        {
-                            var iconFile = iconProp.GetString();
-                            if (iconFile is not null)
-                            {
-                                var full = Path.Combine(extDir, iconFile);
-                                if (File.Exists(full)) { iconPath = full; break; }
-                            }
-                        }
-                    }
-                }
-                if (root.TryGetProperty("homepage_url", out var hp))
-                    homepageUrl = hp.GetString();
-
-                // homepage_url がない場合、Chrome Web Store 拡張なら store URL を構成する
-                if (homepageUrl is null &&
-                    root.TryGetProperty("update_url", out var updateUrl) &&
-                    updateUrl.GetString()?.Contains("clients2.google.com") == true)
-                {
-                    // ext.Id は WebView2 内部 ID のため、フォルダー名（元の CWS ID）を使う
-                    homepageUrl = $"https://chromewebstore.google.com/detail/{Path.GetFileName(extDir)}";
-                }
-            }
-            if (optPage is null) return; // options_ui がない拡張は追加しない
-
-            object btnContent = iconPath is not null
-                ? new Image
-                {
-                    Source = new Microsoft.UI.Xaml.Media.Imaging.BitmapImage(new Uri(iconPath)),
-                    Width = 20, Height = 20
-                }
-                : (object)"🧩";
-
-            var btn = new Button
-            {
-                Content = btnContent,
-                Width   = 32,
-                Height  = 32,
-                Padding = new Thickness(0),
-            };
-            ToolTipService.SetToolTip(btn, string.Format(R.Get("ExtSettings_Format"), name));
-
-            var optPageUrl = $"chrome-extension://{ext.Id}/{optPage}";
-            btn.Click += async (_, _) =>
-            {
-                var optWebView = new WebView2 { Width = 480, MinHeight = 200 };
-
-                Uri.TryCreate(homepageUrl, UriKind.Absolute, out var homepageUri);
-                var linkText = homepageUri?.Host.Contains("chromewebstore.google.com") == true
-                    ? R.Get("ExtSettings_StoreLink")
-                    : R.Get("ExtSettings_Homepage");
-
-                var dlg = new ContentDialog
-                {
-                    Title                = string.Format(R.Get("ExtSettings_Format"), name),
-                    Content              = optWebView,
-                    SecondaryButtonText  = homepageUri is not null ? linkText : null,
-                    CloseButtonText      = R.Get("Button_Close"),
-                    XamlRoot             = Content.XamlRoot
-                };
-
-                // SecondaryButton クリックでブラウザーを開き、ダイアログは閉じない
-                if (homepageUri is not null)
-                    dlg.SecondaryButtonClick += (s, e) =>
-                    {
-                        e.Cancel = true;
-                        _ = Windows.System.Launcher.LaunchUriAsync(homepageUri);
-                    };
-
-                var env = await GetOrCreateProfileEnvAsync("default");
-                await optWebView.EnsureCoreWebView2Async(env);
-                optWebView.Source = new Uri(optPageUrl);
-                await ShowDialogAsync(dlg);
-            };
-
-            // 設定ボタン（末尾）の左隣に挿入
-            int insertIdx = Math.Max(0, RightToolbar.Children.Count - 1);
-            RightToolbar.Children.Insert(insertIdx, btn);
-        }
-
-        private static readonly string LogFilePath = Path.Combine(
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
-            "XTimelineViewer", "error.log");
-
-        private static void LogError(string context, Exception ex)
-        {
-            try
-            {
-                Directory.CreateDirectory(Path.GetDirectoryName(LogFilePath)!);
-                File.AppendAllText(LogFilePath,
-                    $"[{DateTime.Now:yyyy-MM-dd HH:mm:ss}] {context}\n{ex}\n\n");
-            }
-            catch { /* ログ書き込み失敗は無視 */ }
-        }
-
-        private async Task InitWebViewAsync(WebView2 webView, TimelineConfig cfg)
-        {
-            try
-            {
-                var env = await GetOrCreateProfileEnvAsync(cfg.ProfileId);
-                await webView.EnsureCoreWebView2Async(env);
-                webView.CoreWebView2.SourceChanged += (s, e) =>
-                {
-                    bool diverged = !IsOnBaseUrl(webView.CoreWebView2.Source, cfg.Url);
-                    if (diverged)
-                    {
-                        _urlDivergedWebViews.Add(webView);
-                    }
-                    else
-                    {
-                        _urlDivergedWebViews.Remove(webView);
-                        if (Uri.TryCreate(cfg.Url, UriKind.Absolute, out var cfgU) &&
-                            cfgU.AbsolutePath.StartsWith("/home", StringComparison.OrdinalIgnoreCase))
-                            RestartAutoActivateTimer();
-                    }
-                    EvaluateHardReloadPause(webView);
-                };
-                await LoadExtensionsAsync(webView);
-                ApplyThemeToWebViews();
-            }
-            catch (Exception ex)
-            {
-                LogError($"InitWebViewAsync (url={cfg.Url})", ex);
-
-                // XamlRoot が準備できていない場合があるので、ループで待機する
-                for (int i = 0; i < 20 && Content.XamlRoot is null; i++)
-                    await Task.Delay(100);
-
-                if (Content.XamlRoot is not null)
-                {
-                    var dlg = new ContentDialog
-                    {
-                        Title           = R.Get("WebViewInitError_Title"),
-                        Content         = new ScrollViewer
-                        {
-                            MaxHeight = 300,
-                            Content   = new TextBlock
-                            {
-                                Text = $"EdgeDevAppDir: {EdgeDevAppDir}\n\nログ: {LogFilePath}\n\n{ex}",
-                                FontFamily = new FontFamily("Cascadia Mono, Consolas, Courier New"),
-                                FontSize   = 12,
-                                IsTextSelectionEnabled = true,
-                                TextWrapping = TextWrapping.Wrap
-                            }
-                        },
-                        CloseButtonText = R.Get("Button_Close"),
-                        XamlRoot        = Content.XamlRoot
-                    };
-                    await ShowDialogAsync(dlg);
-                }
-                return;
-            }
-
-
-
-            // キーボードショートカット：ブラウザ既定アクセラレータを無効化し JS で代替処理
-            webView.CoreWebView2.Settings.AreBrowserAcceleratorKeysEnabled = false;
-            await webView.CoreWebView2.AddScriptToExecuteOnDocumentCreatedAsync(KeyboardShortcutScript);
-            await webView.CoreWebView2.AddScriptToExecuteOnDocumentCreatedAsync(TimestampInterceptScript);
-            webView.CoreWebView2.WebMessageReceived += (s, e) =>
-                OnWebViewMessageReceived(webView, e.TryGetWebMessageAsString());
-
-            // 外部リンクをシステム既定ブラウザーで開く
-            webView.CoreWebView2.NewWindowRequested += async (s, args) =>
-            {
-                args.Handled = true;
-                await Windows.System.Launcher.LaunchUriAsync(new Uri(args.Uri));
-            };
-
-            webView.CoreWebView2.NavigationStarting += async (s, args) =>
-            {
-                if (!Uri.TryCreate(args.Uri, UriKind.Absolute, out var nav)) return;
-
-                if (Uri.TryCreate(cfg.Url, UriKind.Absolute, out var origin) &&
-                    !nav.Host.Equals(origin.Host, StringComparison.OrdinalIgnoreCase))
-                {
-                    args.Cancel = true;
-                    await Windows.System.Launcher.LaunchUriAsync(nav);
-                    return;
-                }
-
-            };
-
-            webView.CoreWebView2.NavigationCompleted += async (s, args) =>
-            {
-                if (args.IsSuccess)
-                {
-                    await ApplyHideSidebarAsync(webView, cfg.HideSidebar);
-                    await ApplyHideComposeAsync(webView, EffectiveHideCompose(cfg, webView.CoreWebView2.Source));
-                    await ApplyHideListHeaderAsync(webView, cfg.HideListHeader);
-
-                    var tsFlag = _appSettings.OpenTimestampInBrowser ? "true" : "false";
-                    await webView.CoreWebView2.ExecuteScriptAsync(
-                        $"window._xtvOpenTimestampInBrowser = {tsFlag};");
-
-                    // x.com/home の場合だけ新着ポスト自動表示機能を適用する
-                    if (Uri.TryCreate(webView.CoreWebView2.Source, UriKind.Absolute, out var current) &&
-                        current.AbsolutePath.TrimEnd('/').Equals("/home", StringComparison.OrdinalIgnoreCase))
-                    {
-                        await ApplyAutoShowNewPostsAsync(webView, cfg.Url);
-                    }
-                }
-            };
-
-            webView.CoreWebView2.SourceChanged += async (s, args) =>
-            {
-                if (cfg.HideCompose)
-                    await ApplyHideComposeAsync(webView, EffectiveHideCompose(cfg, webView.CoreWebView2.Source));
-                if (cfg.HideListHeader)
-                    await ApplyHideListHeaderAsync(webView, cfg.HideListHeader);
-            };
-
-            webView.Source = new Uri(cfg.Url);
-            StartHardReloadTimer(webView, cfg);
         }
     }
 }
